+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+0.0.6: renamed customVariantSelectors to customVariants and allowed functions as values, to return a rule to use.
+
+0.0.5: added variant chaining, e.g. hover:after:text-red.
+
 0.0.4: revert the un-prefixed classes change from last version, not working.
 
 0.0.3: fixes README formatting. Handles un-prefixed classes better.